Remove ChartInputsContainer
