--- client/src/queryEditor/ChartInputsContainer.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import { handleChartConfigurationFieldsChange } from '../stores/editor-actions';
-import {
-  useSessionQueryResult,
-  useSessionChartType,
-  useSessionChartFields,
-} from '../stores/editor-store';
-import ChartInputs from './ChartInputs';
-
-function ChartInputsContainer() {
-  const queryResult = useSessionQueryResult();
-  const chartType = useSessionChartType();
-  const chartFields = useSessionChartFields();
-
-  return (
-    <ChartInputs
-      chartType={chartType}
-      queryChartConfigurationFields={chartFields}
-      onChartConfigurationFieldsChange={handleChartConfigurationFieldsChange}
-      queryResult={queryResult}
-    />
-  );
-}
-
-export default React.memo(ChartInputsContainer);
